refactor: split project utils into modules
